refactor(hooks): extract pagination params into separate hook
